Add melgan-based WAV output (see notes)
This currently is a hack and appents melgan to the path. I'm going to
move the code into the repo so it can be under proper version control
with this project on a followup commit

Note: I tried to use NVIDIA's WaveGlow, but it was super hard to get
working. The GPU requirements were too much. I couldn't balance between
two GPUs as it out of memoried. I also tried to use the CPU for
inference, but it kept segfaulting.

Melgan appears to work. The initial audio is super robotic/choppy, but
you can sort of tell something is going on.